Updating/Creating Classes
I'm using an old project as a reference to make classes for everything. Seem to have run into an issue. In C#, you can use "import" to include folders and files containing methods and datatypes across your project. Doesn't seem to quite work the same here. Going to have to find the equivalent