So far
Some of the base structs we need
Some descriptions of future stuff.
Before we do anything we need to match the backend counterparts of the types to how they will appear in game. We also need to make sure we can send the right type of data from the backend, to the game logic, to the front end.